+++ ChangeLog.md
@@ -18,6 +18,8 @@ ChangeLog
   `Sabre\DAV\Client::setVerifyPeer`.
 * Removed: `Sabre\DAV\Plugin::getPlugin()` can now no longer return plugins
   based on its class name.
+* Removed: `Sabre\DAVACL\Plugin::getPrincipalByEmail()`.
+
 
 2.1.0-alpha2 (2014-10-23)
 -------------------------